Revamp PerformanceControl
Preview at Patch Set 5:
http://www.youtube.com/watch?v=5lxudfr_9tI

Implements a Navigation Drawer and lets the user switch between tabbed
and drawer navigation.

It comes with options in config.xml, so you can overlay it and disable
the option to let the user choose between tabbed and drawer mode.
You can also toggle if it should use tabbed or drawer navigation
by default.

This above means, we can still use the tabbed only view, without any
options to let the user switch to drawer navigation, if you dont want
that in any case.

Also removing obsolete code, reformatting the whole code and bringing
the code up-to-date.
Updated Support Library (just in case)

Patch Set 2:
Use the action bar HOME as HOME, not to toggle the drawer.
toggle the drawer on menu key press and when pressing the
icon in the action bar on the right.

Patch Set 3 & 4:
Fixes and Cleanups

Patch Set 5:
Add a switch to the action bar, which allows it to switch
between tabbed and drawer mode.

Patch Set 6:
Move the toggle for drawer / tabbed to overflow menu
Add config.xml items for overlaying, now we can select the
default mode and disable user selection, means removing the
toggle from the overflow menu
Also more source cleanups, reformatting and
bringing it up-to-date.

Change-Id: Ib227b0cf52183b19ef22545bd90a6f04fb603d86