Revamp PerformanceControl (WIP)
Preview at Patch Set 1:
http://www.youtube.com/watch?v=JQBljKvmFjs

Implements a Navigation Drawer instead of the ViewPager.
Also removing obsolete code.
Updated Support Library (just in case)

Work in Progress!
All is a mess right now, needs more cleanup.

Patch Set 2:
Use the action bar HOME as HOME, not to toggle the drawer.
toggle the drawer on menu key press and when pressing the
icon in the action bar on the right.

Change-Id: Ib227b0cf52183b19ef22545bd90a6f04fb603d86