Update patch set 1
Patch Set 1:

the top left back button didnt work for me before though?

also we could switch the drawer opening to the menu key, for devices which have one OR set a button to the top right to open it.

so the drawer could be opened via the swipe and the button press and we could go back like always via pressing at the top left.

Patch-set: 1