Update patch set 8
Patch Set 8: -Code-Review

I have just noticed that this is an android-4.3 commit. Is it really meant to be that, or was it a repo start error? 
Blimey, today I have cherry picked it for four different 4.4 device builds, and it works flawlessly on all of them.:p

Patch-set: 8
Reviewer: Gerrit User 1000133 <1000133@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0