Update patch set 8
Patch Set 8:

its 4.3 on purpose.

when i created it, there was no 4.4 branch existing, it got added afterwards (i also meantioned it in the commit message of the first patchset)

in case this gets merged, im sure someone will cherry-pick it to 4.4 or i just resubmit it on the 4.4 branch.

Patch-set: 8