Update patch set 6
Patch Set 6:

Please set the progress bar of the first wakelock to 100%. Then display all others relative to that. Otherwise, there are just empty bars for me. The list looks a bit cluttered, maybe not all information is needed. What about a details dialog? Later, we can add a wakelock blocker to that list, so please provide a clear separation between list entries to make it easier to click on an entry. The wakelock can be blocked directly from the details dialog if you add the dialog.

Patch-set: 6
Label: Code-Review=0